Renaming block, moving frontend JS from `accordion` to `accordions`
