+++ src/blocks/accordion-item/frontend/style.scss
